# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gen-IT API Documentation",
        "description": "API documentation for Gen-IT application",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vicons": {
            "get": {
                "tags": [
                    "Vicons"
                ],
                "summary": "Get all Vicons with pagination",
                "operationId": "22465994ceb1ecebdf568afb37822cb0",
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "_limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "_search",
                        "in": "query",
                        "description": "Search by title or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_by",
                        "in": "query",
                        "description": "Filter by creator",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Vicons with pagination"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vicons"
                ],
                "summary": "Create new vicon",
                "operationId": "6d24fc5e0a8142b8f3e0d2a2d8ff0c75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "desc",
                                    "img",
                                    "time",
                                    "link"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "desc": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "file"
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "link": {
                                        "type": "string"
                                    },
                                    "download": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vicon created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vicon created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vicons/active": {
            "get": {
                "tags": [
                    "Vicons"
                ],
                "summary": "Get all active vicons",
                "description": "Returns list of active (not soft deleted) vicons",
                "operationId": "2c94c3eaafb071a3e8ec19310e4b24e6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vicons/{id}": {
            "get": {
                "tags": [
                    "Vicons"
                ],
                "summary": "Get vicon by ID",
                "operationId": "17008783c8b929d75b4e1c922a243511",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vicon details"
                    },
                    "404": {
                        "description": "Vicon not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vicons"
                ],
                "summary": "Update vicon",
                "operationId": "b53fe1bd67d205caa9618de82d841647",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "desc": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "img": {
                                        "type": "file",
                                        "nullable": true
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "link": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "download": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vicon updated successfully"
                    },
                    "404": {
                        "description": "Vicon not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vicons"
                ],
                "summary": "Delete vicon",
                "operationId": "f7768c0cb0b52b7abc015d9bcf125b6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vicon deleted successfully"
                    },
                    "404": {
                        "description": "Vicon not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}